feat: reusable username input
