+++ client/src/components/Auth/Login.tsx
@@ -1,28 +1,12 @@
-import { useState } from "react";
 import { useNavigate } from "react-router-dom";
-import {
-  Input,
-  Button,
-  InputGroup,
-  Stack,
-  InputLeftElement,
-  chakra,
-  Box,
-  FormControl,
-  InputRightElement,
-} from "@chakra-ui/react";
-import { FaUserAlt, FaLock } from "react-icons/fa";
-import { light_grey_color, white_color } from "../../assets/customColors";
+import { Stack, Box } from "@chakra-ui/react";
+import { white_color } from "../../assets/customColors";
 import BlueButton from "../Button/BlueButton";
-
-const CFaUserAlt = chakra(FaUserAlt);
-const CFaLock = chakra(FaLock);
+import UsernameInput from "./UsernameInput";
+import PasswordInput from "./PasswordInput";
 
 export default function Login() {
-  const [showPassword, setShowPassword] = useState(false);
   const navigate = useNavigate();
-
-  const handleShowClick = () => setShowPassword(!showPassword);
 
   return (
     <Box minW={{ base: "90%", md: "468px" }}>
@@ -33,34 +17,8 @@ export default function Login() {
           backgroundColor={white_color}
           boxShadow="md"
         >
-          <FormControl>
-            <InputGroup>
-              <InputLeftElement
-                pointerEvents="none"
-                children={<CFaUserAlt color={light_grey_color} />}
-              />
-              <Input type="email" placeholder="email address" />
-            </InputGroup>
-          </FormControl>
-          <FormControl>
-            <InputGroup>
-              <InputLeftElement
-                pointerEvents="none"
-                color={light_grey_color}
-                children={<CFaLock color={light_grey_color} />}
-              />
-              <Input
-                type={showPassword ? "text" : "password"}
-                placeholder="Password"
-              />
-              <InputRightElement width="4.5rem">
-                <Button h="1.75rem" size="sm" onClick={handleShowClick}>
-                  {showPassword ? "Hide" : "Show"}
-                </Button>
-              </InputRightElement>
-            </InputGroup>
-            {/* use FormHelperText for message errors*/}
-          </FormControl>
+          <UsernameInput />
+          <PasswordInput />
           <BlueButton
             text="Login"
             onClickAction={() => navigate("/")}
